+++ Source/Animal_Inheritance/BestialityGeneInheritanceDef.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace RJW_BGS
+﻿namespace RJW_BGS
 {
     public class BestialityGeneInheritanceDef
     {