+++ Source/Animal_Inheritance/PatchRJWHediffInsect_Egg.cs
@@ -1,14 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Reflection;
 using System.Reflection.Emit;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using HarmonyLib;
 using rjw;
-using RimWorld;
-using Verse;
 
 namespace RJW_BGS
 {