+++ Source/Animal_Inheritance/RaceGeneDef.cs
@@ -11,11 +11,11 @@ namespace RJW_BGS
     public class RaceGeneDef : Def
     {
         public int priority;
-        public String raceGroup;
+        public string raceGroup; //keeping this for backwards compatibility
+        public List<string> raceGroups; //racegroup, but in list form so multiple can be entered, preference to use this over racegroup 
         public List<string> raceNames;
         public List<string> pawnKindNames;
         public List<BestialityGeneInheritanceDef> genes;
-        //public List<float> genechances;
         public String hybridName;
     }
 }