+++ Source/Genes/GenitaliaSize/Gene_BigMaleGenitalia.cs
@@ -1,20 +1,10 @@
 ﻿namespace RJW_Genes
 {
-    public class Gene_BigMaleGenitalia : RJW_Gene
+    public class Gene_BigMaleGenitalia : Gene_GenitaliaResizingGene
     {
-
-        public override void PostMake()
+        public override void Resize()
         {
-            base.PostMake();
-
-            SizeAdjuster.AdjustAllPenisSizes(pawn,0.5f,1.0f);
-        }
-        
-        public override void PostAdd()
-        {
-            base.PostAdd();
             SizeAdjuster.AdjustAllPenisSizes(pawn, 0.5f, 1.0f);
         }
-
     }
 }